refactor NumberParser
